# FOTD Optimizations: Unify IDs with other fields, implement `__getitems__`, add tests.